gcc: Remove specs CFLAGS from libgcc.mvars
When libgcc is configured it uses the options provided by the top level
CFLAGS_FOR_TARGET variable. These are provided in two sources, both as
part of the CFLAGS for libgcc, but also in the file libgcc.mvars. In
most cases this causes no issues, but in the case of a --specs option,
providing this twice on the command line may cause build failures.

This changes GCC's Makefile to filter out any --specs options when
building libgcc.mvars, so that these are only provided once to libgcc's
configure step.

gcc/ChangeLog:

	* Makefile.in (SPECLESS_GCC_CFLAGS): Define.
  (libgcc.mvars): Use SPECLESS_GCC_CFLAGS instead of GCC_CFLAGS.